changes at kriteria controller: add optimalize

+++ database/migrations/2023_12_04_140019_create_nilai_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     Schema::create('nilais', function (Blueprint $table) {
       $table->foreignUuid('mahasiswa_id');
       $table->foreignUuid('subkriteria_id');
-      $table->float('nilai')->default(0);
+      $table->decimal('nilai')->default(0);
       $table->timestamps();
       $table->softDeletes();
     });
