+++ database/migrations/2024_01_13_034053_create_normalisasis_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('normalisasis', function (Blueprint $table) {
             $table->foreignUuid('mahasiswa_id');
             $table->foreignUuid('kriteria_id');
-            $table->float('nilai')->default(0);
+            $table->decimal('nilai')->default(0);
             $table->timestamps();
             $table->softDeletes();
         });
